drf_api/pyproject.toml
[tool.poetry]
name = "drf_api"
version = "0.1.0"
description = ""
authors = ["sondos-braim <sondos9811@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
